requirements.txt
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
docx2pdf==0.1.8
Flask>=3.1.1
flask-cors==6.0.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
MarkupSafe==3.0.2
narwhals==1.36.0
scipy==1.13.1
numpy==2.2.5
openai==1.76.0
openpyxl
packaging==24.2
pandas==2.2.3
pefile==2024.8.26
pillow==11.2.1
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2025.4
pyperclip
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
setuptools==80.3.1
simpleaudio==1.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
streamlit==1.44.1
streamlit_wavesurfer
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0

